// lib/api/generated/model/getChatbotEmbedSettingResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Uchat API
 * Uchat API Documentation
 * OpenAPI spec version: 1.0
 */
import type { ApiSuccessResponseDto } from "./apiSuccessResponseDto"
import type { ChatbotEmbedSettingResponseDto } from "./chatbotEmbedSettingResponseDto"

export type GetChatbotEmbedSettingResponse = ApiSuccessResponseDto & {
  data?: ChatbotEmbedSettingResponseDto
}